+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- feat: Add `__fileChecksum()` function to `helpers.js`
 
 ## 2.3.0 - 2.3.1